+++ chapter_001/src/test/java/ru/job4j/calculator/ConverterTest.java
@@ -10,28 +10,28 @@ public class ConverterTest {
     @Test
     public void when70RubleToEuroThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToEuro(70);
+        double result = converter.rubleToEuro(70);
         assertThat(result, is(1D));
     }
 
     @Test
     public void when60RubleToDollarThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToDollar(60);
+        double result = converter.rubleToDollar(60);
         assertThat(result, is(1D));
     }
 
     @Test
     public void when1EuroToRubleThen70() {
         Converter converter = new Converter();
-        int result = converter.euroToRuble(1);
+        double result = converter.euroToRuble(1);
         assertThat(result, is(70D));
     }
 
     @Test
     public void when1DollarToRubleThen60() {
         Converter converter = new Converter();
-        int result = converter.dollarToRuble(1);
+        double result = converter.dollarToRuble(1);
         assertThat(result, is(60D));
     }
 }
